django_projects/csv_validator_simple/requirements.txt
cchardet==1.1.3
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
datapackage==0.8.9
Django==1.11.4
djangorestframework==3.6.4
docutils==0.14
et-xmlfile==1.0.1
future==0.16.0
goodtables==1.4.3
gunicorn==19.7.1
idna==2.6
ijson==2.3
isodate==0.5.4
jdcal==1.3
jsonlines==1.2.0
jsonschema==2.6.0
jsontableschema==0.10.1
linear-tsv==1.0.0
openpyxl==2.4.8
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
rfc3986==0.4.1
simpleeval==0.9.5
six==1.10.0
SQLAlchemy==1.1.13
statistics==1.0.3.5
tabulator==1.4.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
